feat: rate && ghost-button-group

+++ scripts/_utils.js
@@ -9,11 +9,13 @@ const path = require("path");
 
 /**
  * Copy a folder and files in it recursively
- * @param {stinrg} source
- * @param {stinrg} target
+ * @param {*} source
+ * @param {*} target
+ * @param {string[]|undefined} excludeList files or folders to escape being copied
  */
-function copyFolderSync(source, target) {
-  // target not existing, create it
+function copyFolderSync(source, target, _excludeList) {
+  const excludeList = _excludeList || [];
+  // create the target folder if it doesn't exist
   if (!fs.existsSync(target)) {
     fs.mkdirSync(target);
   }
@@ -23,26 +25,35 @@ function copyFolderSync(source, target) {
     const sourcePath = path.join(source, file);
     const targetPath = path.join(target, file);
 
-    // copy it recursively if it is a folder
-    if (fs.statSync(sourcePath).isDirectory()) {
-      copyFolderSync(sourcePath, targetPath);
-    } else {
-      // if it is a file, copy it directly
+    // copy it recursively if it is a folder and not in the exclude list
+    if (fs.statSync(sourcePath).isDirectory() && !excludeList.includes(file)) {
+      copyFolderSync(sourcePath, targetPath, excludeList);
+    } else if (!excludeList.includes(file)) {
+      // if it is a file and not in the exclude list, copy it directly
       fs.copyFileSync(sourcePath, targetPath);
     }
   });
 }
 
-function deleteFolderRecursive(folderPath) {
+/**
+ * Delete a folder and files in it recursively
+ * @param {*} folderPath
+ * @param {string[]|undefined} excludeList files or folders to escape being deleted
+ */
+function deleteFolderRecursive(folderPath, _excludeList) {
+  const excludeList = _excludeList || [];
   if (fs.existsSync(folderPath)) {
     fs.readdirSync(folderPath).forEach((file) => {
       const curPath = path.join(folderPath, file);
-      if (fs.lstatSync(curPath).isDirectory()) {
-        // delete subfolders recursively
-        deleteFolderRecursive(curPath);
-      } else {
-        // delete folder files
-        fs.unlinkSync(curPath);
+      if (!excludeList.includes(file)) {
+        // escape the file or folder
+        if (fs.lstatSync(curPath).isDirectory()) {
+          // delete subfolders recursively
+          deleteFolderRecursive(curPath, excludeList);
+        } else {
+          // delete folder files
+          fs.unlinkSync(curPath);
+        }
       }
     });
     // delete empty folders
